ProxyVPN/requirements.txt
mitmproxy==12.1.1
requests==2.32.4
Flask==3.0.2
asyncio  # Python's async networking (built-in for Python 3.4+ so no specific version needed for pip)
# logging  # Built-in logging module
# json  # JSON handling (built-in, but explicitly mentioned)
# os  # Built-in module for file handling
# sys  # System-specific parameters
Flask-CORS==6.0.0
python-dotenv==1.0.1
gunicorn==23.0.0
redis==5.0.7
flask-socketio==5.5.1
pybreaker==1.3.0
tenacity==9.1.2
scikit-learn==1.7.0
joblib==1.4.2
prometheus_client==0.14.1 # Prometheus client library's latest version is 0.14.1 as per PyPI. The search results for "prometheus_client latest stable version" show download links for "prometheus-3.4.1", which refers to the Prometheus server, not the Python client library.
structlog==25.4.0